minor -Wshadow s/*/*/
git-svn-id: https://svn.r-project.org/R/trunk@10928 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/sock.c
@@ -133,7 +133,7 @@ int Sock_listen(int fd, char *cname, int buflen, Sock_error_t perr)
 {
 #if !defined(Unix) || defined(HAVE_BSD_NETWORKING)
   struct sockaddr_in net_client;
-  int len = sizeof(struct sockaddr);
+  size_t len = sizeof(struct sockaddr);
   int retval;
   struct hostent *hostptr;
